Add changelog & small code clean up
