+++ .changes/unreleased/ENHANCEMENTS-494-20241016-113244.yaml
@@ -1,0 +1,5 @@
+kind: ENHANCEMENTS
+body: '* `Agent Pool Deployment` : Add the ability to add labels and annotations to agent deployment.'
+time: 2024-10-16T11:32:44.901754-04:00
+custom:
+    PR: "494"
