+++ api/v1alpha2/agentpool_validation.go
@@ -72,34 +72,6 @@ func (ap *AgentPool) validateSpecAgentToken() field.ErrorList {
 	return allErrs
 }
 
-// Validating labels to ensure key value pairs are not empty
-func validateLabels(labels map[string]string, fldPath *field.Path) field.ErrorList {
-	allErrs := field.ErrorList{}
-	for k, v := range labels {
-		if k == "" {
-			allErrs = append(allErrs, field.Required(fldPath.Child("key"), "key must not be empty"))
-		}
-		if v == "" {
-			allErrs = append(allErrs, field.Required(fldPath.Child("value"), "value must not be empty"))
-		}
-	}
-	return allErrs
-}
-
-// Validate annotations to ensure key value pairs are not empty
-func validateAnnotations(annotations map[string]string, fldPath *field.Path) field.ErrorList {
-	allErrs := field.ErrorList{}
-	for k, v := range annotations {
-		if k == "" {
-			allErrs = append(allErrs, field.Required(fldPath.Child("key"), "key must not be empty"))
-		}
-		if v == "" {
-			allErrs = append(allErrs, field.Required(fldPath.Child("value"), "value must not be empty"))
-		}
-	}
-	return allErrs
-}
-
 // TODO:Validation
 //
 // + Invalid CR cannot be deleted until it is fixed -- need to discuss if we want to do something about it
